feat: Código segregado e com elevação de estado para compartilhamento de dados.
