+++ src/pages/Provisionamento/SearchONU/style.js
@@ -1,3 +1,32 @@
+import styled from "styled-components";
+
 export const Container = styled.div`
 
+`;
+
+export const InputContainer = styled.div`
+    /*
+        ESSE ESTILO CONTROLA O TEXTO E O QUE SEJA INSERIDO PELO USUÁRIO,
+        CARACTERISTICO POR: INFORMAÇÃO EM TEXTO SOBRE O QUE DEVE SER INSERIDO
+        E O CAMPO DE INSERÇÃO PARA O USUÁRIO
+    */
+    
+    display: flex;
+    justify-content: flex-start;
+    width: 80%;
+    margin: 1rem 0;
+
+    @media (max-width: 450px){
+        width: 100%;
+    }
+
+    .text{
+        display: flex;
+        align-items: ${(props) => (props.center ? 'center' : 'end')};
+    }
+
+    .text, .content{
+        width: 50%;
+        padding: 0 2px;
+    }
 `;
